integrated forecast role and collected the needed RL files
